Refactor withdrawal service
