+++ src/main/java/com/payprovider/withdrawal/rest/UserController.java
@@ -1,5 +1,6 @@
 package com.payprovider.withdrawal.rest;
 
+import com.payprovider.withdrawal.exception.TransactionException;
 import com.payprovider.withdrawal.model.User;
 import com.payprovider.withdrawal.repository.UserRepository;
 import com.payprovider.withdrawal.service.UserService;
@@ -20,18 +21,13 @@ public class UserController {
 
     private final UserService userService;
 
-    @Autowired
-    private ApplicationContext context;
-
     @GetMapping("/find-all-users")
     public List<User> findAll() {
-
-        return context.getBean(UserRepository.class).findAll();
+        return userService.findAll();
     }
 
     @GetMapping("/find-user-by-id/{id}")
-    public User findById(@PathVariable Long id) {
-
-        return context.getBean(UserRepository.class).findById(id).orElseThrow();
+    public User findById(@PathVariable Long id) throws TransactionException {
+        return userService.findById(id);
     }
 }
